Avoid exponential work for GHC by rewriting instances for Pair.
The `Rep` of `Pair a` is `(a,a)`, and the generic instances for `HasRep` types
caused a few dictionaries to be constructed twice (for `a` and `a`), leading to
exponentially many constructions for trees. This commit carefully reuses the
same dictionary for uniform pairs.

Also escapes non-ascii characters in prim names in dot files,
apparently needed in recent versions of Graphviz.